+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### [14.1.9](https://kaos.sh/ek/14.1.9)
 
-- **`[terminal]`** Fixed bug with reading data on Windows
+- **`[terminal]`** Fixed bug with reading data on Windows using `Read` and `ReadAnswer`
 - **`[spinner]`** Fixed Windows support
 
 ### [14.1.8](https://kaos.sh/ek/14.1.8)